requirements.txt
requests[socks]>=2.31.0
pystray>=0.19.5
pillow>=10.0.0